api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.45.0
python-multipart==0.0.6
pydantic==2.5.0
pymongo==4.8.0
python-dotenv==1.0.0